Build source distributions only
When we release, we shouldn't be building wheels since we can't build
universal wheels for upload. If we want to build wheels that are
pre-compiled against CPython versions, we'll need to do that with Travis
CI or some other CI/CD distributor and give them the ability to publish
to PyPI for us.

For now, let's stick to source distributions only.