Exclude test files from flake8 checking
We have some static analysis checks that are failing because test methods are missing docstrings. Travis CI had an `allow_failures` feature that allowed us to run static analysis checks and still report a build as passing even if static analysis fails. We're migrating to GitHub Actions and it doesn't have an analogous command, so let's explicitly exclude the two test files that are failing static analysis for now.