Remove benchmark tests from CI
The benchmarks were passing in Travis CI, but we're switching to GitHub Actions and the 3.8 benchmark is consistently failing there. There are likely some external factors coming into play here, so we're going to remove the benchmarks from CI runs. They are still available to be run locally using `tox` or `tox -e <benchmark env name>`.